+++ mask.c
@@ -140,6 +140,7 @@ mask_single_(const uint8_t* __restrict__ v, const uint8_t order)
 static void __attribute__((__nonnull__))
 place_finder_(qrmask_t* self)
 {
+  // TODO: move to lookup
   static const uint8_t finder[7][7] = {
     { 1, 1, 1, 1, 1, 1, 1 },
     { 1, 0, 0, 0, 0, 0, 1 },
@@ -164,6 +165,7 @@ place_finder_(qrmask_t* self)
 static void __attribute__((__nonnull__))
 place_align_(qrmask_t* self)
 {
+  // TODO: move to lookup
   static const uint8_t align[5][5] = {
     { 1, 1, 1, 1, 1 },
     { 1, 0, 0, 0, 1 },
@@ -216,9 +218,11 @@ percentage_penalty_(qrmask_t* self)
 static void __attribute__((__nonnull__))
 module_penalty_(qrmask_t* self)
 {
+  // TODO: move to lookup
   static const uint8_t patright[9] = {
     1, 1, 1, 0, 1, 0, 0, 0, 0
   };
+  // TODO: move to lookup
   static const uint8_t patleft[9] = {
     1, 1, 1, 0, 1, 0, 0, 0, 0
   };
@@ -315,9 +319,11 @@ symbol_order_(const uint8_t version)
 int
 create_qrmask(qrmask_t** self, const uint8_t version, const uint8_t pattern)
 {
+  // TODO: move to lookup
   static const uint16_t qr_basedark[MAX_VERSION] = {
     91u, 112u, 114u, 118u, 122u
   };
+  // TODO: move to lookup
   static const uint16_t qr_offset[MAX_VERSION] = {
     0, 208u, 567u, 1134u, 1941u
   };
@@ -406,6 +412,7 @@ qrmask_penalty(qrmask_t* self)
 void
 qrmask_apply(qrmask_t* self)
 {
+  // TODO: move to lookup
   static const uint16_t maskinfo[NUM_MASKS] = {
     30660u, 29427u, 32170u, 30877u,
     26159u, 25368u, 27713u, 26998u