+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Refactor codebase with `clang-tidy`
 - Fix minimum version selection
 - Fix build issue on MSYS environment
+- Remove deprecated call to `fopen`
 
 ## [1.11.0] - 2025-09-05